+++ packages/expo-dev-menu/android/src/main/java/expo/modules/devmenu/websockets/DevMenuCommandHandlersProvider.kt
@@ -53,6 +53,7 @@ class DevMenuCommandHandlersProvider(
             val activity = instanceManager.currentReactContext?.currentActivity ?: return
             devDelegate.togglePerformanceMonitor(activity)
           }
+          "openJsInspector" -> devDelegate.openJsInspector()
           else -> Log.w("DevMenu", "Unknown command: $command")
         }
       }